% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{export_dataset}
\alias{export_dataset}
\title{Export a dataset.}
\usage{
export_dataset(
  x,
  path = paste0(getwd(), "/", deparse(substitute(x))),
  type = "xlsx"
)
}
\arguments{
\item{x}{A data frame.}

\item{path}{The path to the location on the user's local computer where the data frame should be stored. This parameter is set to the working directory #' by default.}

\item{type}{A character element representing the type of export to perform (options: "xlsx" or "spss"). This parameter is set to "xlsx" by default.}
}
\value{
Returns none.
}
\description{
This function allows users to export a dataset to Excel or SPSS.
}
