% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{get_correlation_coefficient}
\alias{get_correlation_coefficient}
\title{Randomly create a correlation coefficient.}
\usage{
get_correlation_coefficient(size, direction = "p", seed_number = 1)
}
\arguments{
\item{size}{A character element (options: "w" for weak, "s" for small, "m" for moderate or "l" for large).}

\item{direction}{A character element (options: "p" for positive or "n" for negative) This paramater is set to "p" by default.}

\item{seed_number}{A numeric (integer) element that allows a randomly generated effect size to be reproduced. This parameter is set to 1 by default.}
}
\value{
Returns a numeric element between -0.70 and 0.70.
}
\description{
This function randomly creates a correlation coefficient.
}
