% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.r
\name{get_dataset}
\alias{get_dataset}
\title{Create a demo dataset}
\usage{
get_dataset(n = 1000)
}
\arguments{
\item{n}{A numberic (integer) element representing the number of participants (rows) in the dataset. This parameter is set to 1000 by default.}
}
\value{
Returns a data frame.
}
\description{
This function creates a demo dataset.
}
